✨ mentorMainPage 서버 연동 완료

+++ lib/src/data/dto/response/mentor/mentor_response.dart
@@ -18,7 +18,7 @@ class MentorResponse with _$MentorResponse {
     int? bookmarkCount,
     int? reviewCount,
     int? mentoringCount,
-    int? reviewScore,
+    double? reviewScore,
     required MemberResponse member
   }) = _MentorResponse;
